Change fiscal year to integer field

+++ opre_ops/ops_site/models.py
@@ -73,7 +73,7 @@ class CANAmount(models.Model):
     information by fiscal year for a given CAN
     """
     can = models.ForeignKey(CANInfo, on_delete=models.PROTECT)
-    fiscal_year = models.DecimalField(max_digits=4, decimal_places=0)
+    fiscal_year = models.IntegerField()
     amount_available = models.DecimalField(max_digits=12, decimal_places=2)
     amount_budgeted = models.DecimalField(max_digits=12, decimal_places=2)
     additional_amount_anticipated = models.DecimalField(max_digits=12, decimal_places=2)
